fix(zvm): wrap loadw/storew array addresses at 16 bits; un-vacuum praxix test (SQ-0239)
loadw/loadb/storew/storeb computed the array address in 32-bit space
(a + 2*index as u32) instead of wrapping at 16 bits as the reference
interpreter does. A negative word-index (praxix passes 0xFFFF = -1) then
produced a huge out-of-bounds address (a + 0x1FFFE) that memory-faulted and
halted the machine, where frotz wraps to a-2 (in-bounds). Compute the address
with u16 wrapping arithmetic.

The praxix regression test was vacuous — it fed empty input, but praxix does
not auto-run (unlike CZECH), so it never executed and its "no failures"
assertion was trivially true. It now drives the core opcode/undo/table groups
and asserts each reports "Passed." without the interpreter faulting (which
directly guards this fix via the "Array loads and stores" group).

praxix additionally surfaced three separate pre-existing gaps, filed as
SQ-0240 (output stream 3 UTF-8 vs ZSCII), SQ-0241 (@scan_table branch), and
SQ-0242 (1.1 true colour); those groups are excluded from the gate until fixed.

Completes: SQ-0239

+++ crates/zvm/src/cpu/exec.rs
@@ -465,16 +465,21 @@ impl Machine {
                 objects::insert_obj(&mut self.mem, a, b);
                 StepResult::Continue
             }
-            // 0x0F loadw — load word from array: result = mem[a + 2*b]
+            // 0x0F loadw — load word from array: result = mem[a + 2*b].
+            // The address is computed in 16-bit space and wraps at 0xFFFF (as the
+            // reference interpreter does), so a negative word-index (e.g. b=0xFFFF
+            // meaning -1) reads the word *before* the array rather than a huge OOB
+            // address. Regression-guarded by praxix's "Array loads and stores"
+            // group (crates/zvm/tests/regression.rs).
             0x0F => {
-                let addr = (a as u32).wrapping_add(2u32.wrapping_mul(b as u32));
+                let addr = a.wrapping_add(2u16.wrapping_mul(b)) as u32;
                 let result = self.mem.read_word(addr);
                 self.do_store(store, result);
                 StepResult::Continue
             }
-            // 0x10 loadb — load byte from array: result = mem[a + b]
+            // 0x10 loadb — load byte from array: result = mem[a + b] (16-bit wrap).
             0x10 => {
-                let addr = (a as u32).wrapping_add(b as u32);
+                let addr = a.wrapping_add(b) as u32;
                 let result = self.mem.read_byte(addr) as u16;
                 self.do_store(store, result);
                 StepResult::Continue
@@ -860,20 +865,22 @@ impl Machine {
                 call_routine(&mut self.state, &mut self.mem, packed, args, store);
                 StepResult::Continue
             }
-            // 0x01 storew — store word: mem[ops[0] + 2*ops[1]] = ops[2]
+            // 0x01 storew — store word: mem[ops[0] + 2*ops[1]] = ops[2].
+            // Address computed in 16-bit space (wraps at 0xFFFF), matching the
+            // reference interpreter, so negative indices address within memory.
             0x01 => {
-                let array = ops.first().copied().unwrap_or(0) as u32;
-                let index = ops.get(1).copied().unwrap_or(0) as u32;
+                let array = ops.first().copied().unwrap_or(0);
+                let index = ops.get(1).copied().unwrap_or(0);
                 let val   = ops.get(2).copied().unwrap_or(0);
-                self.mem.write_word(array.wrapping_add(2u32.wrapping_mul(index)), val);
+                self.mem.write_word(array.wrapping_add(2u16.wrapping_mul(index)) as u32, val);
                 StepResult::Continue
             }
-            // 0x02 storeb — store byte: mem[ops[0] + ops[1]] = ops[2] & 0xFF
+            // 0x02 storeb — store byte: mem[ops[0] + ops[1]] = ops[2] & 0xFF (16-bit wrap)
             0x02 => {
-                let array = ops.first().copied().unwrap_or(0) as u32;
-                let index = ops.get(1).copied().unwrap_or(0) as u32;
+                let array = ops.first().copied().unwrap_or(0);
+                let index = ops.get(1).copied().unwrap_or(0);
                 let val   = (ops.get(2).copied().unwrap_or(0) & 0xFF) as u8;
-                self.mem.write_byte(array.wrapping_add(index), val);
+                self.mem.write_byte(array.wrapping_add(index) as u32, val);
                 StepResult::Continue
             }
             // 0x03 put_prop — set property ops[1] of object ops[0] to value ops[2]
